Impl crate restructure (#30)
* move arithmetic impls into submodule

* move cmp impls into submodule

* move logic impls into submodule

* move morphism into submodule

* move pipelines into submodule

* move io into submodule

* clean up intrinsics module
